Refactor the LEDC driver to reduce code duplication
